# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Blockchain Manager API
  version: "0.6.0"
  x-curl: true
  description: |
    블록체인 매니저는 사내의 별도 서비스로, 온체인 거래(노드 연동)를 담당한다.
    호출 쪽 백엔드(Service·Admin)는 이 HTTP API 로 계정·주소·잔액·거래를 다루고,
    온체인 상태 변경은 메시지 큐 이벤트로 받는다.

    아래 규약은 **모든 엔드포인트에 공통** 적용된다.

    ## 응답 형식

    성공·목록·에러 모두 같은 구조로 돌려준다. `meta.requestId` 로 요청을 추적한다. 스키마 이름은 단건이 `<타입>Response`, 목록이 `<타입>ListResponse` 다.

    단일 리소스:

    ```json
    {
      "data": {
        "accountType": "CUSTOMER",
        "ref": "000123",
        "accountId": "acct_01H8X"
      },
      "meta": {
        "requestId": "3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f"
      }
    }
    ```

    페이지네이션 목록:

    ```json
    {
      "data": [
        { "txId": "tx_9f2a", "status": "FINALIZED", "amount": "1.5" }
      ],
      "meta": {
        "requestId": "3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f"
      },
      "pagination": {
        "nextCursor": "eyJsYXN0IjoxNzUxMzM2MDAwMDAwfQ",
        "hasMore": true
      }
    }
    ```

    에러:

    ```json
    {
      "error": {
        "code": "ACCOUNT_NOT_FOUND",
        "message": "account not found"
      },
      "meta": {
        "requestId": "3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f"
      }
    }
    ```

    ## 데이터 포맷

    - **시각** — ISO 8601, UTC, 밀리초. 예: `2026-07-13T04:05:06.789Z`
    - **금액** — 문자열(decimal). 예: `"1.5"`. float 가 아니라 decimal 로 파싱한다.
    - **필드명** — camelCase (`externalTxId` · `numOfConfirmations`)
    - **요청 추적** — 모든 응답에 `meta.requestId`
...

    ## 에러 코드

...

...

...

...

...

...

...
servers:
...
paths:
...
  /transactions:
    post:
      tags: [Transactions]
      summary: 출금 제출
      description: |
        출금(또는 내부 이체)을 제출한다. 응답은 벤더 tx id(`txId`)이고 상태 진행은 큐 이벤트로 따라간다(Events).

        - `externalTxId` 가 멱등 키다. **같은 키 + 같은 내용**을 다시 보내면 처음의 `txId` 를 돌려주므로 **재시도가 안전**하다.
        - 같은 키인데 **내용이 다르면** `409` 다.
        - 제출한 건은 `GET /transactions/external/{externalTxId}` 로 찾는다 — 출금은 출금 풀 vault 에서 나가 계정별 목록에는 없다.

        **"같은 내용"의 범위** — 자금이 어디서 어디로 얼마나 움직이는지를 규정하는 값만 본다:
        `from.type` · `from.accountId` · `to.type` · `to` 의 식별값(`address`·`accountId`·`walletId` 중 채워진 하나) · `network` · `symbol` · `amount`.

        - `note` 와 `travelRule` 은 **비교하지 않는다.** 메모는 자금 이동을 바꾸지 않고, 트래블룰 산출물은 다시 만들면 값이 달라질 수 있어 정당한 재시도를 막게 된다.
        - `amount` 는 금액으로 비교한다 — `"1.50"` 과 `"1.5"` 는 같다.
        - 나머지는 **문자 그대로** 비교한다. 주소·네트워크·심볼은 대소문자를 바꾸지 않으므로, 재시도할 때는 처음 보낸 문자열을 그대로 보내야 한다.

        **응답을 못 받았을 때** — `5xx` 나 타임아웃은 제출이 나갔는지 알 수 없다는 뜻이지 실패했다는 뜻이 아니다.
        같은 요청을 그대로 다시 보내면 되고(멱등), 확인만 하려면 `GET /transactions/external/{externalTxId}` 를 쓴다.

        **같은 키를 동시에 보냈을 때** — 앞선 요청이 아직 처리 중이면 `503 SUBMIT_IN_PROGRESS` 와 `Retry-After` 가 온다.
        중복 제출이 아니라 **아직 결과를 모른다**는 뜻이라, 그 시간만큼 기다렸다 같은 요청을 그대로 다시 보내면 된다.
      operationId: submitTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionRequest' }
      responses:
        '202':
          description: 접수됨(제출)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubmitResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '404': { $ref: '#/components/responses/AccountNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { $ref: '#/components/responses/RelayRejected' }
        '503': { $ref: '#/components/responses/SubmitInProgress' }
  /transactions/external/{externalTxId}:
    parameters:
      - name: externalTxId
        in: path
        required: true
        schema: { type: string, maxLength: 128 }
        description: 제출할 때 실은 우리 요청 키
        example: wd-260713-0042
    get:
      tags: [Transactions]
      summary: 우리 요청 키로 거래 조회
      description: |
        `externalTxId` 로 제출한 건을 찾는다. 출금은 고객 계정이 아니라 **출금 풀 vault 에서 나가므로** 계정별 목록 조회로는 찾을 수 없다 — 호출 쪽이 자기 출금을 아는 유일한 키가 `externalTxId` 라 이 경로가 필요하다.

        제출 응답을 못 받았을 때의 확인, 그리고 대사에서 우리 기록과 벤더 기록을 잇는 데 쓴다.
      operationId: transactionByExternalTxId
      responses:
        '200':
          description: 조회 결과
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransferResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /transactions/{txId}:
    get:
      tags: [Transactions]
      summary: 거래 단건 조회
      description: 벤더 tx id(`txId`)로 거래 1건을 조회한다. `txId` 는 출금 제출 응답이나 큐 이벤트에서 얻는다.
      operationId: transactionOf
      parameters:
        - name: txId
          in: path
          required: true
          schema: { type: string }
          description: 벤더 tx id
          example: tx_9f2a
      responses:
        '200':
          description: 거래
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransferResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /accounts/{accountId}/transactions:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags: [Transactions]
      summary: 거래 목록 조회
      description: |
        거래 이력을 **거래 시각(createdAt) 기준**으로 조회한다 — 기본 최신순, `order=asc` 면 과거→최신. 기간(`after`/`before`)·상태로 좁히고 커서로 페이지네이션한다.
        `order=asc` + `before` 생략 조합이면 마지막 `nextCursor` 를 보관했다가 재요청해 새로 쌓인 내역만 이어받는 증분 폴링이 된다.
        상태 변경 실시간 감지는 이 목록이 아니라 이벤트 큐가 담당한다(매니저의 웹훅 감지와 별개).

        **커서는 매니저가 발급한다** — 벤더 커서를 그대로 넘겨주지 않는다. 최초 요청의 필터와 정렬, 그리고 이어받을 위치를
        매니저가 토큰에 담으므로, 마지막 페이지에서도 `nextCursor` 가 채워지고 그 값으로 증분 폴링이 성립한다.
        `hasMore=false` 는 "지금 시점에 더 없다"는 뜻이지 커서가 끝났다는 뜻이 아니다.

        **매핑되지 않은 자산의 거래는 목록에서 빠진다** — 우리가 등록하지 않은 벤더 자산이 섞여 오면 그 건만 제외하고 나머지를 돌려준다.
        빠뜨린 건은 조용히 버리지 않고 운영 알림으로 올린다(등록 누락이면 고쳐야 할 설정이다).
      operationId: transactionsOf
      parameters:
        - name: after
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: |
            시작 시각 — 거래 시각(createdAt) 기준 (ISO 8601 UTC).
            **첫 요청(`cursor` 없음)에는 필수**고, 없으면 `400 VALIDATION_FAILED` 다.
            `cursor` 가 있으면 조회 조건이 토큰에 들어 있어 이 값은 무시되므로 생략한다.
          example: "2026-07-01T00:00:00.000Z"
        - name: before
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: 종료 시각 — 거래 시각(createdAt) 기준 (ISO 8601 UTC). 생략하면 상한 없음 — 증분 폴링(`order=asc`) 조회는 생략한다.
          example: "2026-07-13T00:00:00.000Z"
        - name: order
          in: query
          required: false
          schema: { type: string, enum: [asc, desc], default: desc }
          description: 정렬 방향 — 거래 시각(createdAt) 기준. 기본 desc(최신순). 마지막 커서를 보관해 새 내역을 이어받는 증분 폴링은 `asc` 조회에서만 성립한다.
          example: desc
        - name: status
          in: query
          required: false
          schema: { $ref: '#/components/schemas/TxStatus' }
          description: 상태 필터 (선택)
          example: FINALIZED
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 200 }
          description: 페이지 크기 — 기본 200, 최대 500 (벤더 한도). 1 미만이거나 500 초과면 `400 VALIDATION_FAILED`.
          example: 200
        - name: cursor
          in: query
          required: false
          schema: { type: string }
          description: 다음 위치 커서 — 이전 응답의 `pagination.nextCursor` 를 그대로 넣는다. 불투명 토큰이라 직접 만들거나 해석하지 않는다. 첫 요청엔 생략. cursor 가 있으면 조회 조건은 토큰이 우선이라 함께 보낸 `after`/`before`·`status`·`order`·`limit` 는 무시된다.
          example: eyJsYXN0IjoxNzUxMzM2MDAwMDAwfQ
      responses:
        '200':
          description: 거래 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransferListResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '404': { $ref: '#/components/responses/AccountNotFound' }
components:
  parameters:
...
    AccountId:
      name: accountId
      in: path
      required: true
      schema: { type: string, maxLength: 64 }
      description: 매니저가 돌려준 vault 핸들 (DB ext_acnt_id = vaultAccountId)
      example: acct_01H8X
  schemas:
...
    Meta:
      type: object
      properties:
        requestId:
          type: string
          description: 요청 추적 id (모든 응답에 포함)
          example: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f
      required: [requestId]
    Pagination:
      type: object
      properties:
        nextCursor:
          type: string
          description: 다음 위치 커서 (불투명 토큰) — 다음 요청 `cursor` 로 그대로 전달. 마지막 페이지에서도 항상 채워지며, `order=asc` 조회면 보관해 뒀다가 이후 새로 쌓인 내역을 이어받는 시작점(증분 폴링)으로 쓴다.
          example: eyJsYXN0IjoxNzUxMzM2MDAwMDAwfQ
        hasMore: { type: boolean, description: 지금 이어받을 다음 페이지가 있는지 — false 면 현재 시점 마지막 페이지, example: true }
      required: [nextCursor, hasMore]
    ErrorBody:
      type: object
      properties:
        code:
          type: string
          description: 에러 코드 (API Conventions 표 참조)
          example: ACCOUNT_NOT_FOUND
        message:
          type: string
          description: 사람이 읽는 설명 — 분기 판단은 `code` 로 한다
          example: account not found
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
        meta: { $ref: '#/components/schemas/Meta' }
      required: [error, meta]
...
    Transfer:
      type: object
      description: 거래 1건. 요청의 `from`/`to`(TransferPeer)는 여기선 확정된 온체인 주소 문자열로 나온다.
      properties:
        txId: { type: string, description: 벤더 tx id }
        txHash:
          type: [string, "null"]
          description: 온체인 거래해시 — 전파 후 채워짐
        externalTxId:
          type: [string, "null"]
          description: 우리 요청 키
        network: { type: string, description: 네트워크 코드, example: ETHEREUM }
        symbol: { type: string, description: 토큰 심볼, example: USDC }
        amount: { type: string, description: 금액(문자열) }
        from:
          type: [string, "null"]
          description: |
            발신 (확정 온체인 주소). **`SUBMITTED` 구간에는 비어 있을 수 있다** — 거래가 체인에 오르기 전에는
            벤더가 주소를 확정하지 않는다. 키는 항상 있고 값만 `null` 이다.
        to:
          type: [string, "null"]
          description: 목적지 (확정 온체인 주소). `from` 과 같은 이유로 `SUBMITTED` 구간에는 비어 있을 수 있다.
        status: { $ref: '#/components/schemas/TxStatus' }
        numOfConfirmations: { type: integer, description: 누적 컨펌 수 }
        createdAt: { type: string, format: date-time, description: 거래 생성 시각 (목록 정렬·기간 필터 기준) }
        lastUpdated: { type: string, format: date-time, description: 마지막 상태 변경 시각 }
      required: [txId, network, symbol, amount, from, to, status, numOfConfirmations, createdAt, lastUpdated]
      example:
        txId: tx_9f2a
        txHash: "0x4e1d...ab"
        externalTxId: wd-260713-0042
        network: ETHEREUM
        symbol: USDC
        amount: "1.5"
        from: "0xA1...C9"
        to: "0x9f...E2"
        status: FINALIZED
        numOfConfirmations: 12
        createdAt: "2026-07-13T04:05:06.789Z"
        lastUpdated: "2026-07-13T04:06:10.120Z"
...
    TxStatus:
      type: string
      enum: [SUBMITTED, CONFIRMED, FINALIZED, REJECTED, FAILED]
      description: 공통 상태 다섯 — 매니저와 호출 쪽 사이의 계약 어휘 (벤더 원어와 구분).
      x-enumDescriptions:
        SUBMITTED: 제출 — 체인 미등장
        CONFIRMED: 체인 등장·컨펌 누적 (미확정 — 확정은 FINALIZED)
        FINALIZED: 확정 — DCCP 임계 도달 (체인 finality 아님 — reorg 시 FAILED 전이 존재)
        REJECTED: 거부·차단 (일시적)
        FAILED: 영구 실패
...
    SubmitResult:
      type: object
      properties:
        txId: { type: string, description: 벤더 tx id, example: tx_9f2a }
      required: [txId]
...
    TransactionRequest:
      type: object
      properties:
        externalTxId:
          type: string
          maxLength: 128
          description: 우리 요청 키 — 승인 완료된 출금 지시 1건과 1:1. 재제출 중복 차단·완료 대응. 초과 시 `400 VALIDATION_FAILED`.
          example: wd-260713-0042
        from:
          allOf:
            - { $ref: '#/components/schemas/TransferPeer' }
            - properties: { type: { const: ACCOUNT } }
          description: 보내는 쪽 — type=ACCOUNT 만 허용
        to:
          allOf: [{ $ref: '#/components/schemas/TransferPeer' }]
          description: 목적지
        network: { type: string, description: 네트워크 코드, example: ETHEREUM }
        symbol: { type: string, description: 토큰 심볼, example: USDC }
        amount:
          type: string
          pattern: '^\d{1,18}(\.\d{1,18})?$'
          description: |
            금액(문자열 · 부동소수 금지). **0보다 커야 하고**, 정수부 최대 18자리 · 소수부 최대 18자리다.
            반올림 없이 그대로 보관할 수 있는 범위이며, 벗어나면 `400 VALIDATION_FAILED` 다.
            부호·지수 표기(`1e-3`)·앞뒤 공백은 허용하지 않는다.
            멱등 비교는 금액으로 하므로 `"1.50"` 과 `"1.5"` 는 같은 요청이다.
          example: "1.5"
        note:
          type: [string, "null"]
          description: 벤더 거래 기록 메모
        travelRule:
          description: 트래블룰 게이트가 만든 암호화 산출물 — 해외(Notabene) 출금만 싣고, 국내(VerifyVASP)·개인지갑은 null
          oneOf:
            - { $ref: '#/components/schemas/TravelRule' }
            - { type: "null" }
      required: [externalTxId, from, to, network, symbol, amount]
      example:
        externalTxId: wd-260713-0042
        from: { type: ACCOUNT, accountId: acct_pool_02 }
        to: { type: ADDRESS, address: "0x9f...E2" }
        network: ETHEREUM
        symbol: USDC
        amount: "1.5"
        note: null
        travelRule: null
    TransferPeer:
      type: object
      description: 벤더 TransferPeerPath 대응. from·to 공통. type 에 따라 필요한 식별 필드가 정해진다.
      properties:
        type: { $ref: '#/components/schemas/PeerType' }
        address:
          type: [string, "null"]
          description: 온체인 주소 (type=ADDRESS 일 때 필수)
        accountId:
          type: [string, "null"]
          description: 우리 계정 (type=ACCOUNT 일 때 필수)
        walletId:
          type: [string, "null"]
          description: 사전 등록 지갑 id (type=WHITELISTED 일 때 필수)
      required: [type]
      allOf:
        - if: { properties: { type: { const: ADDRESS } } }
          then: { required: [address] }
        - if: { properties: { type: { const: ACCOUNT } } }
          then: { required: [accountId] }
        - if: { properties: { type: { const: WHITELISTED } } }
          then: { required: [walletId] }
    PeerType:
      type: string
      enum: [ADDRESS, ACCOUNT, WHITELISTED]
      x-enumDescriptions:
        ADDRESS: 온체인 주소 (외부 출금 → ONE_TIME_ADDRESS)
        ACCOUNT: 우리 계정 (내부 이동 → VAULT_ACCOUNT)
        WHITELISTED: 사전 등록 지갑 (→ EXTERNAL_WALLET)
    TravelRule:
      type: object
      description: |
        트래블룰 게이트가 만든 **암호화 산출물**이다. 이 API(매니저)는 운반만 하고
        내용을 파싱하지 않으므로, 여기서는 내부 구조를 펼치지 않고 불투명한 객체로 둔다.

        - 실제 구조의 기준은 **IVMS101 표준 + 트래블룰 솔루션 스펙**(게이트 쪽 문서)이다.
        - 시나리오별로 실림 여부가 다르다 — 해외(Notabene)=메시지 있음, 국내(VerifyVASP)·개인지갑=없음(`null`).
        - 컴플라이언스가 내보내는 `travelRuleMessage`(암호화 문자열)를 호출 쪽이 이 필드로 실어 보낸다 — 정확한 형태는 구현 때 확정.
      additionalProperties: true
...
    TransferResponse:
      type: object
      properties:
        data: { $ref: '#/components/schemas/Transfer' }
        meta: { $ref: '#/components/schemas/Meta' }
      required: [data, meta]
    TransferListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Transfer' }
        meta: { $ref: '#/components/schemas/Meta' }
        pagination: { $ref: '#/components/schemas/Pagination' }
      required: [data, meta, pagination]
    SubmitResponse:
      type: object
      properties:
        data: { $ref: '#/components/schemas/SubmitResult' }
        meta: { $ref: '#/components/schemas/Meta' }
      required: [data, meta]
  responses:
    ValidationFailed:
      description: 요청 검증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: VALIDATION_FAILED, message: "amount must be a decimal string" }
            meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: NOT_FOUND, message: "transaction not found" }
            meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
    AccountNotFound:
      description: 계정 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: ACCOUNT_NOT_FOUND, message: "account not found" }
            meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
    Conflict:
      description: 상태·멱등 충돌
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: CONFLICT, message: "externalTxId already used" }
            meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
    SubmitInProgress:
      description: |
        같은 `externalTxId` 의 앞선 제출이 처리 중이다. 중복 제출이 아니라 아직 결과를 모른다는 뜻이고,
        `Retry-After` 초 뒤에 같은 요청을 그대로 다시 보내면 된다.
      headers:
        Retry-After:
          description: 다시 보내기까지 기다릴 초
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: SUBMIT_IN_PROGRESS, message: "submission for this externalTxId is in progress" }
            meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
    RelayRejected:
      description: relay 가 전송을 대지 못함·거절 (대납 구성)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: RELAY_REJECTED, message: "relay refused to sponsor gas" }
            meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }